requirements.txt
PyAudio
numpy
pillow